feat(model) decrement method (#9169)
