+++ lib/model.js
@@ -2923,30 +2923,33 @@ class Model {
 
   /**
    * Increment the value of one or more columns. This is done in the database, which means it does not use the values currently stored on the Instance. The increment is done using a
-   * ```sql
-   * SET column = column + X WHERE foo = 'bar'
-   * ```
-   * query. To get the correct value after an increment into the Instance you should do a reload.
+   * ``` SET column = column + X WHERE foo = 'bar' ``` query. To get the correct value after an increment into the Instance you should do a reload.
    *
-   *```js
-  * Model.increment('number', { where: { foo: 'bar' }) // increment number by 1
-  * Model.increment(['number', 'count'], { by: 2, where: { foo: 'bar' } }) // increment number and count by 2
-  * Model.increment({ answer: 42, tries: -1}, { by: 2, where: { foo: 'bar' } }) // increment answer by 42, and decrement tries by 1.
-  *                                                        // `by` is ignored, since each column has its own value
-  * ```
-  *
-  * @see {@link Model#reload}
-  * @param {String|Array|Object} fields If a string is provided, that column is incremented by the value of `by` given in options. If an array is provided, the same is true for each column. If and object is provided, each column is incremented by the value given.
-  * @param {Object} options
-  * @param {Object} options.where
-  * @param {Integer} [options.by=1] The number to increment by
-  * @param {Boolean} [options.silent=false] If true, the updatedAt timestamp will not be updated.
-  * @param {Function} [options.logging=false] A function that gets executed while running the query to log the sql.
-  * @param {Transaction} [options.transaction]
-  * @param  {String}       [options.searchPath=DEFAULT] An optional parameter to specify the schema search_path (Postgres only)
-  *
-  * @return {Promise<this>}
-  */
+   * ```js
+   * // increment number by 1
+   * Model.increment('number', { where: { foo: 'bar' });
+   *
+   * // increment number and count by 2
+   * Model.increment(['number', 'count'], { by: 2, where: { foo: 'bar' } });
+   *
+   * // increment answer by 42, and decrement tries by 1.
+   * // `by` is ignored, since each column has its own value
+   * Model.increment({ answer: 42, tries: -1}, { by: 2, where: { foo: 'bar' } });
+   * ```
+   *
+   * @see {@link Model#reload}
+   *
+   * @param {String|Array|Object} fields If a string is provided, that column is incremented by the value of `by` given in options. If an array is provided, the same is true for each column. If and object is provided, each column is incremented by the value given.
+   * @param {Object} options
+   * @param {Object} options.where
+   * @param {Integer} [options.by=1] The number to increment by
+   * @param {Boolean} [options.silent=false] If true, the updatedAt timestamp will not be updated.
+   * @param {Function} [options.logging=false] A function that gets executed while running the query to log the sql.
+   * @param {Transaction} [options.transaction]
+   * @param  {String}       [options.searchPath=DEFAULT] An optional parameter to specify the schema search_path (Postgres only)
+   *
+   * @return {Promise<this>}
+   */
   static increment(fields, options) {
     options = options || {};
 
@@ -3007,6 +3010,36 @@ class Model {
 
       return [affectedRows];
     });
+  }
+
+  /**
+   * Decrement the value of one or more columns. This is done in the database, which means it does not use the values currently stored on the Instance. The decrement is done using a
+   * ```sql SET column = column - X WHERE foo = 'bar'``` query. To get the correct value after a decrement into the Instance you should do a reload.
+   *
+   *```js
+   // decrement number by 1
+  * Model.decrement('number', { where: { foo: 'bar' });
+  *
+  * // decrement number and count by 2
+  * Model.decrement(['number', 'count'], { by: 2, where: { foo: 'bar' } });
+  *
+  * // decrement answer by 42, and decrement tries by -1.
+  * // `by` is ignored, since each column has its own value
+  * Model.decrement({ answer: 42, tries: -1}, { by: 2, where: { foo: 'bar' } });
+  * ```
+  *
+  * @see {@link Model#increment}
+  * @see {@link Model#reload}
+  * @since 4.36.0
+
+  * @return {Promise<this>}
+  */
+  static decrement(fields, options) {
+    options = _.defaults({ increment: false }, options, {
+      by: 1
+    });
+
+    return this.increment(fields, options);
   }
 
   static _optionsMustContainWhere(options) {
